refactor: change progressPercentage type from BigDecimal to Double

+++ src/main/java/com/EdumentumBackend/EdumentumBackend/dtos/course/CourseDetailResponseDto.java
@@ -3,7 +3,6 @@ package com.EdumentumBackend.EdumentumBackend.dtos.course;
 import com.EdumentumBackend.EdumentumBackend.enums.EnrollmentStatus;
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -18,6 +17,6 @@ public class CourseDetailResponseDto {
     private List<ResourceResponseDto> resources;
     private Boolean isEnrolled;
     private EnrollmentStatus enrollmentStatus;
-    private BigDecimal progressPercentage;
+    private Double progressPercentage;
     private RatingResponseDto userRating;
 }
